<refactor> player객체 삭제 후 car객체 내부 move함수 사용
