--- src/main/kotlin/data/generator/CarRacingGamePlayerGenerator.kt
@@ -1,11 +1,0 @@
-package data.generator
-
-import data.Car
-import data.CarRacingGamePlayer
-
-class CarRacingGamePlayerGenerator {
-
-    fun generateCarRacers(cars: List<Car>, numberOfTry: Int) = cars.map { car ->
-        CarRacingGamePlayer(car, numberOfTry)
-    }
-}
