+++ src/main/kotlin/domain/CarRacingGameController.kt
@@ -2,25 +2,21 @@ package domain
 
 import data.Car
 import data.CarPath
-import data.CarRacingGamePlayer
 import data.generator.CarGenerator
-import data.generator.CarRacingGamePlayerGenerator
 import util.RESULT
 import view.InputView
 import view.OutputView
 
 class CarRacingGameController(
     private val carRacingGame: CarRacingGame = CarRacingGame(),
-    private val carGenerator: CarGenerator = CarGenerator(),
-    private val carRacingGamePlayerGenerator: CarRacingGamePlayerGenerator = CarRacingGamePlayerGenerator()
+    private val carGenerator: CarGenerator = CarGenerator()
 ) {
 
     fun startGame() {
         try {
             val cars = initCars()
             val numberOfTry = initNumberOfTry()
-            val players = initPlayers(cars, numberOfTry)
-            val carsPath = carRacingGame.startDriving(players)
+            val carsPath = carRacingGame.startDriving(cars, numberOfTry)
 
             showPath(carsPath, numberOfTry)
             showWinner(cars)
@@ -42,9 +38,6 @@ class CarRacingGameController(
 
         return numberOfTry.toInt()
     }
-
-    private fun initPlayers(cars: List<Car>, numberOfTry: Int): List<CarRacingGamePlayer> =
-        carRacingGamePlayerGenerator.generateCarRacers(cars, numberOfTry)
 
     private fun showWinner(cars: List<Car>) {
         OutputView.printWinner(carRacingGame.decideWinner(cars))
